refactor: 채팅 기능 useChatRoom 훅으로 분리
