+++ frontend/src/components/ChatRoom/ChatRoom.jsx
@@ -1,52 +1,11 @@
-import { useEffect, useRef, useState } from 'react';
 import styles from './ChatRoom.module.css';
-import { chatClient } from '../../utils/chat/chatClient';
 import { useParams } from 'react-router-dom';
 import SendIcon from '/src/assets/icons/send.svg?react';
-import useBoundStore from '../../store';
+import useChatRoom from '../../hooks/chat/useChatRoom';
 
-const USER_ID = crypto.getRandomValues(new Uint32Array(1))[0];
-
-// TODO: 채팅 훅 분리
 export default function ChatRoom() {
   const { lectureId } = useParams();
-  const client = chatClient;
-  const [messages, setMessages] = useState([]);
-  const userName = useBoundStore((state) => state.userName) ?? '익명';
-  const inputRef = useRef(null);
-  const handleSubmit = (e) => {
-    e.preventDefault();
-    const text = inputRef.current.value;
-
-    if (!text) {
-      return;
-    }
-
-    chatClient.publish({
-      destination: `/pub/message/${lectureId}`,
-      body: JSON.stringify({
-        userId: USER_ID,
-        name: userName,
-        content: text,
-      }),
-    });
-    inputRef.current.value = '';
-  };
-
-  useEffect(() => {
-    client.onConnect = () => {
-      client.subscribe(`/sub/channel/${lectureId}`, (response) => {
-        const data = JSON.parse(response.body);
-        const { content: message, name } = data;
-        setMessages((prev) => [...prev, { id: prev.length, text: message, isMine: USER_ID === data.userId, name }]);
-      });
-    };
-    client.activate();
-
-    return () => {
-      client.deactivate();
-    };
-  }, [client, lectureId]);
+  const { messages, handleSubmit, inputRef } = useChatRoom(lectureId);
 
   return (
     <section className={styles.room}>
